Feat: 소비내역 분석, 무비티아이 로직 변경

See merge request s11-fintech-finance-sub1/S11P21A404!127

+++ backend/moobeing/src/main/java/com/im/moobeing/domain/deal/dto/CategoryPercentDto.java
@@ -9,9 +9,9 @@ import lombok.Getter;
 @AllArgsConstructor
 public class CategoryPercentDto {
     private String label;
-    private Double percent;
+    private Integer percent;
 
-    public static CategoryPercentDto of(String label, Double percent) {
+    public static CategoryPercentDto of(String label, Integer percent) {
         return CategoryPercentDto.builder()
                                  .label(label)
                                  .percent(percent)
